Make sure /home is mounted with noexec,nodev
Previously, .xinitrc created various symlinks from /home/webc into /tmp, which
was mounted with noexec for extra security. This removes the symlink stuff and
just makes sure /home is mounted with noexec instead. If /home is not a
separate filesystem yet, it is mounted as a new tmpfs (preserving the old
contents).