delete .git-fixups/000-chown-root.sh
The chown is not needed with git-fs (and shouldn't be needed when the chroot is
created and edited as root either). Also, this chown breaks when running it at
boot time with git-fs, since it also tries to chown the contents of the ISO.

The cmod on tmp is also in a different fixup script, so it can be dropped as well.