delete some dpkg-related backup files
Seems these files are used as backups and are thus not really needed when we
have git for backups. Keeping these around does cause them to show changes that
are unrelated to the package currently being installed (for example), which
pollutes git commits. This is caused because the backup files lag behind one
change.

Conflicts:
	var/cache/debconf/config.dat-old
	var/lib/dpkg/available-old
	var/lib/dpkg/diversions-old
	var/lib/dpkg/status-old