Re-organisation due to noexec limitation on /home
Appeared to work before because /home was remounted with "noexec" _after_ .xinitrc was started on the first X invocation and
xinit does a fallback to running /bin/sh when .xinitrc fails due to noexec on the second X run.

Despite the "noexec" flag applied on home "/home/webc/webc.sh" is placed in /home because it is run by the "webc" user and seems like the most natural place for it.
http://serverfault.com/questions/72356/how-useful-is-mounting-tmp-noexec

Now webc.sh (originally .xinitrc) is the main "loop" of Webconverger, which in turn triggers live-config.sh/updates.sh to do tasks of a higher privilege.