make the install script work again with git-fs
It now installs a git repository, extlinux configuration files, a kernel and an
initrd, instead of a full root filesystem onto the hard disk.

The extlinux setup is still a bit crude, it just copies stuff from /boot.
Ideally, the extlinux configuration would be regenerated from scratch (also to
facilitate a failover setup when changing kernels / upgrading the root
filesystem).