in the install script, use set -e for errorhandling again
The previous approach, using a ERR trap but not set -e could not automatically
handle failures inside functions. But using an ERR trap together with set -e
did not give the expected results either. To improve this, we now use set -e
again to make the script exit on any error and an EXIT trap to show a message
on a failed install. To prevent a succesfull install from triggering the
failure handler, we unset it after a succesful install.