Support fetch_revision and update_revision
fetch_revision is the revision we should fetch from the git server (it
can only be a branch or tag name, since we can't fetch a sha directly).

update_revision is optional and can be used to specify a sha of a
revision to update to (but it must be "included" in the fetch triggered
by fetch_revision). If no update_revision is given, the revision
specified by fetch_revision is used.