Completely generate extlinux config on install
Instead of using the Debian default extlinux config and heavily
modifying it, we now generate a simple config from scratch. This config
has just a single entry with no prompt, so the installed version always
boots unatteneded.

Also, the kernel is installed using the same code used for upgrades, so
instead of copying the kernel and initrd from the current rootfs, they
are extracted from git directly.