Support upgrades in the installed version
This reuses the same code for building the bootloader config that is
used during the install, which should make sure the bootloader config
stays consistent.

Note that since a few commits, upgrades were broken for installed
versions (possibly to the point of breaking a working install). Now they
work again completely.