Remove / ignore locales and includes
This removes all files in /usr/include and sets up dpkg to not unpack and git
to ignore files in /usr/include and /usr/locales.

This also modifies a bunch of dpkg .list files to remove the locale files, but
it seems these are not all of them (haven't found a way to make dpkg re-check
and update all of its .list files yet).

Since these locale files are only used for translations of software we're not
exposing in webconverger (Firefox uses its own translation mechanism), we're
leaving them out to save space. Include files should not be needed for similar
reasons.