Fix failed_install handler
It was originally written to check $? to see if the last command failed
or if it was called because of a normal exit, but since the handler is
uninstalled just before the normal exit, this was no longer needed.
However, the $? argument was no longer passed to the failed_install
handler, but it did still check it, causing the handler to never work at
all.