Install the locales package
This makes locale stuff available to applications like they expect, preventing
segfaults in e.g. libfuse when printing help options and probably other
sideeffects.

To save space, only the files required for the en_US@UTF-8 locale are
installed. This is the only locale actually generated and also set as the
default locale.