Backport live-build changes to fix using the overlay filesystem
Since kernel 3.18, Debian no longer includes aufs support in their
kernels. Instead, the overlay filesystem is included in upstream kernels
now and should be used instead.

live-build already supported overlayfs, which was an external kernel
patch before 3.18. In 3.18, overlayfs was merged into the mainline
kernel and for compatibility reasons, it was renamed to "overlay".

This commit backports overlay-related changes from live-boot 5.0~a4-1.
In particular, it changes overlayfs to overlay in the relevant places
and it adds a few mount options that are now required.

This commit intentionally does not regenerate the initrd, since these
changes are only needed when upgrading the kernel (which will regenerate
the initrd anyway).