Automatically use overlay when the selected union filesystem is not available
This allows keeping the default at aufs, to work with kernels < 3.18,
while automatically switching to overlay for 3.18 and above (which no
longer support aufs, but have overlay).

This commit intentionally does not regenerate the initrd, since these
changes are only needed when upgrading the kernel (which will regenerate
the initrd anyway).