Let dpkg exclude directories too
Before, the content of a number of directories was excluded, but the
directories itself were not. This caused empty directories to be created by
dpkg, which were not committed to git (due to being empty and ignored).

When dpkg installs a package, it creates a .list file containing all files in
the package, including the ones ignored. However, when dpkg creates a directory
when installing a package, it will remove that directory from the .list files
of all other packages (presumably since, if the directory wasn't present, it
must be erronously listed in the .list files).

This behaviour previously meant that, when installing a package containing
ignored files, the .list files from other packages were sometimes also
modified, leading to messy commits and history. Excluding the directories as
well should prevent the "remove stuff from .list files" behaviour of dpkg being
triggered, leading to cleaner history.