Prevent systemd from forwarding to syslog
Systemd already stores log entries in memory and offers a
syslog-compatible interface, so there is no point in also storing
messages in syslog as well.

Upstream systemd disabled forwarding by default in version 216, but the
Debian package kept the default as enabled.

This should prevent slowdowns and perhaps also a lockup during shutdown
due to a lot of failed syslog forward attempts. The symptom is a lot of
messages on the console like this:

	[  287.191533] systemd-journald[416]: Failed to forward syslog message: Connection refused