Add net.ifnames=0 to the kernel commandline
This causes the kernel to stick to the old behaviour of using eth0 and
wlan0 for networking devices, instead of the newer device naming based
on how they are connected in the system.

In the long term, it might be good to make the switch and configure
network devices more dynamically, but to keep things working with the
fairly static /etc/network/interfaces for now, add this option.